+++ tests/testthat/test-tab_cols.R
@@ -127,8 +127,8 @@ test_that("cols_omit drops column names - stable", {
 test_that("cols_omit drops column names - longtable", {
   a <- stable_long(stdata(), cols_omit = FALSE)
   b <- stable_long(stdata(), cols_omit = TRUE)
-  expect_match(a[1:7], "STUDY &", all = FALSE, fixed = TRUE)
-  expect_match(a[1:7], " & CRCL &", all = FALSE, fixed = TRUE)
+  expect_match(a[1:20], "STUDY &", all = FALSE, fixed = TRUE)
+  expect_match(a[1:20], " & CRCL &", all = FALSE, fixed = TRUE)
   expect_false(any(grepl("STUDY", b)))
   expect_false(any(grepl("CRCL", b)))
 })