+++ NEWS.md
@@ -1,3 +1,19 @@
+# pmtables 0.4.1
+
+- `colgroup()` (and `st_span()`) gains an `align` argument to position
+  the spanner title on left or right in addition to the center (default)
+  #260, #261.
+
+- `rowpanel()` (and `st_panel()`) gains `jut`argument to push non-panel
+  table contents to the right relative to the panel header row so that
+  contents under the panel header are indented #251, #253. 
+
+- Panel header rows are now modified so that the header row stays
+  with the first non-header row for longtable output #252, #253.
+
+- Consistent `BQL` / `BLQ` handling for column titles and table 
+  notes for `pt_data_inventory()` #254, #255. 
+
 # pmtables 0.4.0
 
 - Add `cols_omit` option to omit column header data  (#213)
@@ -11,7 +27,7 @@
 - Add `hline` argument to `rowgroup()` constructor to make the horizontal line
   above the panel data optional (#215)
 
-- Refactor `pt_data_inventory()` to calculate percent BLQ using denomninator
+- Refactor `pt_data_inventory()` to calculate percent BLQ using denominator
   that is the sum of the number of observations BLQ and non-BLQ / non-missing
   (#221, #222)